refactor: split AuthContext and AuthProvider into separate files
Fix Fast Refresh warning by separating the context and component
into their own modules.
