+++ client/src/context/AuthContext.jsx
@@ -1,39 +1,3 @@
-import { createContext, useState, useEffect, useCallback } from 'react';
+import { createContext } from 'react';
 
 export const AuthContext = createContext();
-
-/* eslint-disable react-refresh/only-export-components */
-export const AuthProvider = ({ children }) => {
-  const [token, setToken] = useState(localStorage.getItem('token'));
-  const [user, setUser] = useState(JSON.parse(localStorage.getItem('user') || 'null'));
-
-  const login = useCallback((newToken, newUser) => {
-    localStorage.setItem('token', newToken);
-    localStorage.setItem('user', JSON.stringify(newUser));
-    setToken(newToken);
-    setUser(newUser);
-  }, []);
-
-  const logout = useCallback(() => {
-    localStorage.removeItem('token');
-    localStorage.removeItem('user');
-    setToken(null);
-    setUser(null);
-  }, []);
-
-  // Sync state if localStorage changes in other tabs (optional but good practice)
-  useEffect(() => {
-    const handleStorageChange = () => {
-      setToken(localStorage.getItem('token'));
-      setUser(JSON.parse(localStorage.getItem('user') || 'null'));
-    };
-    window.addEventListener('storage', handleStorageChange);
-    return () => window.removeEventListener('storage', handleStorageChange);
-  }, []);
-
-  return (
-    <AuthContext.Provider value={{ token, user, isLoggedIn: !!token, login, logout }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
